(refactor(mailbridge)): split scanner and runtime test files
- Extract Outlook COM reflection helpers into `OutlookComHelpers` and update the architecture diagrams to reflect the new boundary
- Split oversized runtime test partials into dedicated pipe and calendar files, remove the ad hoc Pester coverage script, and keep all affected files under the 500-line policy limit
- Record remediation, QA, Windows acceptance, and final audit artifacts, including corrected feature-completion evidence and updated plan/spec checkoffs for issue #12

Refs: #12
