+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/ac-status.2026-04-10T17-35.md
@@ -1,0 +1,62 @@
+# Acceptance Criteria Reconciliation — QA Gate Evidence
+
+Timestamp: 2026-04-10T17-35
+
+## Issue Checklist
+
+Source: `docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/issue.md`
+Total AC items: 10
+Checked off (delivered): 10
+Remaining (unchecked): 0
+
+All 10 acceptance criteria in `issue.md` are checked `[x]`.
+
+## Spec Checklist
+
+Source: `docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/spec.md`
+Section: Definition of Done
+Total AC items: 7
+Checked off (delivered): 7
+Remaining (unchecked): 0
+
+All 7 Definition of Done items in `spec.md` are checked `[x]`. The previously unchecked item ("Behavior matches acceptance criteria in the documented Windows interactive-session environment") was verified and checked off during this reconciliation based on P6-T1 and P6-T2 evidence.
+
+Note: `spec.md` also contains 3 "Seeded Test Conditions" items that remain unchecked. These are advisory test planning items, not formal acceptance criteria. They describe test coverage categories rather than deliverable behavior. The corresponding test coverage exists (Phase 5 tests cover DTO/ID helpers, RPC validation, privacy shaping, stale-cache, client behavior, and script assertions), but the seeded conditions are left unchecked because they are planning-stage guidance markers, not AC checkboxes.
+
+## User Story Checklist
+
+Source: `docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/user-story.md`
+Section: Acceptance Criteria
+Total AC items: 10
+Checked off (delivered): 10
+Remaining (unchecked): 0
+
+All 10 acceptance criteria in `user-story.md` are checked `[x]`.
+
+## FrameworkTargetVerification
+
+- All 4 project files contain `<TargetFramework>net10.0-windows</TargetFramework>`
+- Installed bridge runtimeconfig: tfm=net10.0, Microsoft.NETCore.App 10.0.0
+- Installed client runtimeconfig: tfm=net10.0, Microsoft.NETCore.App 10.0.0
+- Evidence: `framework-targets.2026-04-10T17-35.md`
+
+## Evidence References
+
+### QA Gate Evidence (Phase 7, 2026-04-10T17-22 and 2026-04-10T17-35)
+- `csharpier-check.2026-04-10T17-22.md` — CSharpier formatting pass
+- `msbuild-analyzers.2026-04-10T17-22.md` — MSBuild analyzer pass
+- `msbuild-nullable.2026-04-10T17-22.md` — MSBuild nullable pass
+- `coverage.2026-04-10T17-22.md` — C# test+coverage pass
+- `coverage-summary.2026-04-10T17-22.md` — C# coverage summary
+- `coverage-thresholds.2026-04-10T17-22.md` — C# coverage thresholds PASS
+- `powershell-format.2026-04-10T17-22.md` — PowerShell format pass
+- `powershell-analyze.2026-04-10T17-35.md` — PowerShell analyze pass
+- `powershell-test.2026-04-10T17-35.md` — PowerShell test pass (19 tests, 0 failures)
+- `powershell-coverage.2026-04-10T17-35.md` — PowerShell coverage (78.7%)
+- `powershell-coverage-summary.2026-04-10T17-35.md` — PowerShell coverage summary
+- `powershell-coverage-thresholds.2026-04-10T17-35.md` — PowerShell coverage thresholds FAIL (baseline measurement artifact)
+- `framework-targets.2026-04-10T17-35.md` — Framework targets PASS
+
+### Windows Acceptance Evidence (Phase 6)
+- `windows-acceptance.2026-04-10T17-22.md` — Automated suites A,B,C,D,F passed, .NET 10 target confirmed
+- `windows-operator-validation.2026-04-10T17-22.md` — Operator validation: PrimaryInteractiveSession=true, OpenClawSvcPipeConnect=true, NetworkDenyVerified=true
