+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/coverage-summary.2026-04-10T17-22.md
@@ -1,0 +1,10 @@
+# C# Coverage Summary — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Command:** PowerShell coverage summary analysis (git merge-base diff + Cobertura line-rate parsing)
+- **EXIT_CODE:** 0
+- **Output Summary:**
+  - PostChangeOverallLineCoverage: 89.4
+  - ChangedOrNewLineCoverage: 100
+  - NewProductionCoverage: 100
+  - CoverageReportPath: C:\Users\DanMoisan\repos\open-claw-bridge\TestResults\qa-csharp\coverage.cobertura.xml
