+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/coverage.2026-04-10T17-22.md
@@ -1,0 +1,6 @@
+# C# Tests With Coverage — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Command:** `dotnet-coverage collect "dotnet test OpenClaw.MailBridge.sln -c Debug --no-build --logger trx --results-directory TestResults/qa-csharp" -f cobertura -o 'TestResults/qa-csharp/coverage.cobertura.xml'`
+- **EXIT_CODE:** 0
+- **Output Summary:** Test summary: total: 87, failed: 0, succeeded: 86, skipped: 1, duration: 11.0s. OverallLineCoverage: 89.4%. TestAssemblyPath: C:\Users\DanMoisan\repos\open-claw-bridge\tests\OpenClaw.MailBridge.Tests\bin\Debug\net10.0-windows\OpenClaw.MailBridge.Tests.dll. Coverage report: TestResults/qa-csharp/coverage.cobertura.xml.
