+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/csharpier-check.2026-04-10T17-22.md
@@ -1,0 +1,6 @@
+# CSharpier Check — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Command:** `csharpier format .`
+- **EXIT_CODE:** 0
+- **Output Summary:** Formatted 37 files in 217ms. No files changed (verified via `git diff --name-only`). All C# source files are already correctly formatted.
