+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/csharpier-check.2026-04-10T23-00.md
@@ -1,0 +1,6 @@
+# CSharpier Check — QA Gate
+
+Timestamp: 2026-04-10T23-00
+Command: `csharpier check .`
+EXIT_CODE: 0
+Output Summary: Checked 40 files in 346ms. No files needed formatting changes. (First pass found 2 files needing formatting — `OutlookComHelpers.cs` and `OutlookScanner.cs` — which were fixed with `csharpier format .`. Second pass clean.)
