+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/msbuild-analyzers.2026-04-10T17-22.md
@@ -1,0 +1,6 @@
+# Analyzer Build — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Command:** `dotnet build OpenClaw.MailBridge.sln -c Debug /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true`
+- **EXIT_CODE:** 0
+- **Output Summary:** Build succeeded in 3.9s. All four projects compiled with zero warnings and zero errors under .NET analyzers and code-style enforcement. Projects: OpenClaw.MailBridge.Contracts (net10.0-windows), OpenClaw.MailBridge.Client (net10.0-windows), OpenClaw.MailBridge (net10.0-windows), OpenClaw.MailBridge.Tests (net10.0-windows).
