+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/msbuild-nullable.2026-04-10T17-22.md
@@ -1,0 +1,6 @@
+# Nullable Build — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Command:** `dotnet build OpenClaw.MailBridge.sln -c Debug /p:Nullable=enable /p:TreatWarningsAsErrors=true`
+- **EXIT_CODE:** 0
+- **Output Summary:** Build succeeded in 1.1s. All four projects compiled with nullable reference types enabled and TreatWarningsAsErrors=true, producing zero warnings and zero errors. Projects: OpenClaw.MailBridge.Contracts (net10.0-windows), OpenClaw.MailBridge.Client (net10.0-windows), OpenClaw.MailBridge (net10.0-windows), OpenClaw.MailBridge.Tests (net10.0-windows).
