+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/powershell-coverage-summary.2026-04-10T17-35.md
@@ -1,0 +1,13 @@
+# PowerShell Coverage Summary — QA Gate Evidence
+
+Timestamp: 2026-04-10T17-35
+Command: pwsh -NoLogo -NoProfile -ExecutionPolicy Bypass (PowerShell coverage summary — git merge-base delta analysis against TestResults/qa-powershell/coverage.json)
+EXIT_CODE: 0
+Output Summary: PostChangeOverallLineCoverage: 78.7, ChangedOrNewLineCoverage: 77.67, NewProductionCoverage: 100.0. Coverage artifact: TestResults/qa-powershell/coverage.json.
+
+## Parsed Results
+
+- PostChangeOverallLineCoverage: 78.7
+- ChangedOrNewLineCoverage: 77.67
+- NewProductionCoverage: 100
+- CoverageArtifactPath: TestResults/qa-powershell/coverage.json
