+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/powershell-format.2026-04-10T17-22.md
@@ -1,0 +1,6 @@
+# PowerShell Format — QA Gate
+
+- **Timestamp:** 2026-04-10T17:22
+- **Tool:** `mcp_drmcopilotext_run_poshqc_format`
+- **EXIT_CODE:** 0
+- **Output Summary:** PoshQC format ran successfully against the workspace. No PowerShell files were changed (verified via `git diff --name-only -- '*.ps1' '*.psm1' '*.psd1'`). All PowerShell source files are already correctly formatted.
