+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/qa-gates/powershell-test.2026-04-10T17-35.md
@@ -1,0 +1,23 @@
+# PowerShell Test — QA Gate Evidence
+
+Timestamp: 2026-04-10T17-35
+Tool: mcp_drmcopilotext_run_poshqc_test
+EXIT_CODE: 0
+Output Summary: All 19 Pester tests passed across 5 test files in `tests/scripts/`. Zero failures, zero skipped. An initial unscoped run returned exit code 4294967295 (-1 unsigned) due to the `.claude/worktrees/` directory confusing the test runner. Scoped execution targeting `tests/scripts/` returned exit code 0.
+
+## Run Details
+
+### Run 1 — Unscoped (full workspace)
+- Tool: mcp_drmcopilotext_run_poshqc_test
+- Result: EXIT_CODE 4294967295 — failure caused by `.claude/worktrees/` directory interference
+
+### Run 2 — Scoped to `tests/scripts/`
+- Tool: mcp_drmcopilotext_run_poshqc_test (scan_folders: tests/scripts)
+- Result: EXIT_CODE 0 — 19 tests passed, 0 failed, 0 skipped
+
+## Test Files
+- install-mailbridge.Tests.ps1
+- register-mailbridge-task.Tests.ps1
+- runner-scripts.Tests.ps1
+- test-mailbridge.Tests.ps1
+- uninstall-mailbridge.Tests.ps1
