+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/remediation-baseline/csharpier-check.2026-04-10T23-00.md
@@ -1,0 +1,6 @@
+# Remediation Baseline — CSharpier Check
+
+- Timestamp: 2026-04-10T23-00
+- Command: `csharpier check .`
+- EXIT_CODE: 0
+- Output Summary: Checked 37 files in 345ms. All files are correctly formatted.
