+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/remediation-baseline/dotnet-test-coverage.2026-04-10T23-00.md
@@ -1,0 +1,6 @@
+# Remediation Baseline — dotnet test with Coverage
+
+- Timestamp: 2026-04-10T23-00
+- Command: `dotnet test OpenClaw.MailBridge.sln -c Debug --collect:"XPlat Code Coverage"`
+- EXIT_CODE: 0
+- Output Summary: 87 total tests (86 passed, 1 skipped, 0 failed). Line coverage: 83.8% (1270/1515 lines). Branch coverage: 67.6%. Coverage report: `tests/OpenClaw.MailBridge.Tests/TestResults/254e9a54-924a-46d2-926a-ffff7c086e98/coverage.cobertura.xml`.
