+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/remediation-baseline/msbuild-analyzers.2026-04-10T23-00.md
@@ -1,0 +1,6 @@
+# Remediation Baseline — MSBuild Analyzers
+
+- Timestamp: 2026-04-10T23-00
+- Command: `dotnet build OpenClaw.MailBridge.sln -c Debug /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true`
+- EXIT_CODE: 0
+- Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). All four projects compiled (Contracts, Client, MailBridge, Tests).
