+++ docs/features/active/2026-04-07-finish-outlook-mail-bridge-12/evidence/remediation-baseline/msbuild-nullable.2026-04-10T23-00.md
@@ -1,0 +1,6 @@
+# Remediation Baseline — MSBuild Nullable Analysis
+
+- Timestamp: 2026-04-10T23-00
+- Command: `dotnet build OpenClaw.MailBridge.sln -c Debug /p:Nullable=enable /p:TreatWarningsAsErrors=true`
+- EXIT_CODE: 0
+- Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). Nullable analysis passed for all four projects.
